+++ src/main/java/com/epam/zt/testing/action/UserCategoryDeleteAction.java
@@ -7,19 +7,26 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 public class UserCategoryDeleteAction implements Action {
-    ActionResult userCategory = new ActionResult("userCategory");
+    ActionResult userCategory = new ActionResult("userCategory?delete=1", true);
+    ActionResult back = new ActionResult("userCategory");
 
     @Override
     public ActionResult execute(HttpServletRequest req, HttpServletResponse resp) {
         User user = (User) req.getSession().getAttribute("foundUser");
+        User currentUser = (User) req.getSession().getAttribute("user");
+        int idUser = user.getId();
+        int idCurrent = currentUser.getId();
+        if (idUser == idCurrent) {
+            req.setAttribute("deleteError", 1);
+            return back;
+        }
         String roleName = user.getRole().getName();
         if (roleName.equals("user")) {
             UserService.deleteUser(user);
         } else {
             UserService.deleteChild(user);
         }
-        req.setAttribute("delete", "User is deleted");
-        req.setAttribute("found", "");
+        req.getSession().setAttribute("found", "");
         return userCategory;
     }
 }
